Use a global 'NC' object
