+++ Makefile
@@ -2,6 +2,7 @@ include libs/NickelHook/NickelHook.mk
 
 override LIBRARY  := libnickelclock.so
 override SOURCES  += src/nickelclock.cc 
+override MOCS     += src/nickelclock.h
 override CFLAGS   += -Wall -Wextra -Werror
 override CXXFLAGS += -Wall -Wextra -Werror -Wno-missing-field-initializers
